refactor: decompose AppEventHandlerScope session subscriptions (#3878)
Split the 605-line session subscription part into event-family parts
(observe, civilian, naval/army, diplomacy, debug), extract
DebugCommandSessionHandler, and add repo.app_event_handler_part_size CI gate.

Refs #3878
